requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
Django==5.1.7
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
mpmath==1.3.0
nudenet==3.4.2
numpy==2.2.4
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
packaging==24.2
prompt_toolkit==3.0.50
protobuf==6.30.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.24.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==23.0.0
whitenoise==6.9.0